Move MySQL query code out of pollJournals into common
